+++ testcafe/page-object/admin/project/create-project-modal.ts
@@ -1,0 +1,10 @@
+import { t, Selector } from 'testcafe';
+
+import { formControlInput } from '../../../testUtil';
+
+export namespace createProjectModal {
+    export async function getName(): Promise<string> {
+        const name = await formControlInput('name').value;
+        return name || '';
+    }
+}